term: ignore first OSC title change sequence on Windows
ConPTY emits a sequence that sets the title to the name of the
program that is initially launched into it.

This commit tries to ignore that sequence in that circumstance,
so that the logic in b5d156c282a01f2f1cb10aee98655a4a2c2c71e5
can more dynamically set the tab title.